[Refactor]: 인수 테스트 관련 테스트코드 계층 구조 변경
