+++ src/test/java/kuit/subway/utils/fixtures/StationFixtures.java
@@ -1,8 +1,8 @@
-package kuit.subway.station.acceptance.fixture;
+package kuit.subway.utils.fixtures;
 
 import kuit.subway.station.dto.request.StationCreateRequest;
 
-public class SubwayFixture {
+public class StationFixtures {
 
     public static StationCreateRequest 지하철역_생성_요청(String name) {
         return StationCreateRequest.builder()
